Allow copying of team name
