+++ src/shared/Elements/Clipboard.tsx
@@ -32,16 +32,29 @@ class ClipboardComponent extends React.Component<IClipboardProps, {}> {
         >
           {value}
         </Box>
-        <Box ml={'10px'}>
-          <Image
-            imgHeight={'15px'}
-            src={CopyImage}
-            ref={(element) => {
-              this.copy = element;
-            }}
-            style={{ cursor: 'pointer' }}
-          />
+        <Box ml={'18px'} mt={'2px'}>
+          <div className="copy-link">
+            <Image
+              imgHeight={'20px'}
+              src={CopyImage}
+              ref={(element) => {
+                this.copy = element;
+              }}
+            />
+          </div>
         </Box>
+
+        <style jsx>{`
+          .copy-link {
+            cursor: pointer;
+            filter: invert(42%) sepia(65%) saturate(398%) hue-rotate(196deg) brightness(83%) contrast(93%);
+            transition: filter 0.25s ease-in;
+          }
+
+          .copy-link:hover {
+            filter: invert(28%) sepia(100%) saturate(3848%) hue-rotate(209deg) brightness(102%) contrast(108%);
+          }
+        `}</style>
       </Flex>
     );
   }
